[CARBONDATA-4255] Prohibit DROP DATABASE when databaselocation is inconsistent
Why is this PR needed?
When carbon.storelocation and spark.sql.warehouse.dir are configured to different values, the databaselocation maybe inconsistent. When DROP DATABASE command is executed, maybe both location (carbon dblcation and hive dblocation) will be cleared, which may confuses the users

What changes were proposed in this PR?
Drop database is prohibited when database locaton is inconsistent