[CARBONDATA-4305] Support Carbondata Streamer tool for incremental fetch and merge from kafka and DFS Sources
Why is this PR needed?
In the current Carbondata CDC solution, if any user wants to integrate it with a streaming source then he
need to write a separate spark application to capture changes which is an overhead. We should be able to
incrementally capture the data changes from primary databases and should be able to incrementally ingest
the same in the data lake so that the overall latency decreases. The former is taken care of using
log-based CDC systems like Maxwell and Debezium. Here is a solution for the second aspect using Apache Carbondata.

What changes were proposed in this PR?
Carbondata streamer tool is a spark streaming application which enables users to incrementally ingest data
from various sources, like Kafka(standard pipeline would be like MYSQL => debezium => (kafka + Schema registry) => Carbondata Streamer tool)
and DFS into their data lakes. The tool comes with out-of-the-box support for almost all types of schema
evolution use cases. With the streamer tool only add column support is given with drop column and
other schema changes capability in line in the upcoming days. Please refer to design document for
more details about usage and working of the tool.

This closes #4235