[CARBONDATA-4286] Fixed measure comparator
Why is this PR needed?
Select query on a table with and filter condition returns an empty result
while valid data present in the table.

Root cause: Currently when we are building the min-max index at block level,
that time we are using unsafe byte comparator for either dimension or measure
column which returns incorrect result for measure columns.

What changes were proposed in this PR?
We should use different comparators for dimensions and measure columns which
we are already doing at time of writing the min-max index at blocklet level.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #4217