[CARBONDATA-4228] [CARBONDATA-4203] Fixed update/delete after alter add segment
Why is this PR needed?
Deleted records are reappearing or updated records are showing old values in select
queries. It is because after horizontal compaction delete delta file for the external
segment is written to the default path which is Fact\part0\segment_x\ while if the
segment is an external segment then delete delta file should be written to the path
where the segment is present.

What changes were proposed in this PR?
After delete/update operation on the segment, horizontal compaction will be triggered.
Now after horizontal compaction for external segments, the delete delta file will be
written to the segment path at the place of the default path.

This closes #4220