[CARBONDATA-4293] Make Table created without external keyword as Transactional table
Why is this PR needed?
Currently, when you create a table with location( without external keyword) in cluster,
the corresponding table is created as transactional table. If External keyword is
present, then it is created as non-transactional table. This scenario is not handled
in local mode.

What changes were proposed in this PR?
Made changes, to check if external keyword is present or not. If present, then make
the corresponding table as transactional table.

This closes #4221