[CARBONDATA-4243] Fixed si with column meta cache on same column
Why is this PR needed?
Currently, the select query fails when table contains SI and column_meta_cache
on the same columns with to date() UDF. This is happening because pushdownfilters
is null in CarbonDataSourceScanHelper and it is causing null pointer exception.

What changes were proposed in this PR?
At place of passing null value for pushdownfilters in CarbonDataSourceScan.doCanonicalize passed Seq.empty.

This closes #4225