[CARBONDATA-4296]: schema evolution, enforcement and deduplication utilities added
Why is this PR needed?
This PR adds schema enforcement, schema evolution and deduplication capabilities for
carbondata streamer tool specifically. For the existing IUD scenarios, some work
needs to be done to handle it completely, for example -
1. passing default values and storing them in table properties.

Changes proposed for the phase 2 -
1. Handling delete use cases with upsert operation/command itself. Right now we
consider update as delete + insert. With the new streamer tool, it is possible that
user sets upsert as the operation type and incoming stream has delete records as well.
What changes were proposed in this PR?

Configs and utility methods are added for the following use cases -
1. Schema enforcement
2. Schema evolution - add column, delete column, data type change scenario
3. Deduplicate the incoming dataset against incoming dataset itself. This is useful
in scenarios where incoming stream of data has multiple updates for the same record
and we want to pick the latest.
4. Deduplicate the incoming dataset against existing target dataset. This is useful
when operation type is set as INSERT and user does not want to insert duplicate records.

This closes #4227