[CARBONDATA-4318]Improve load overwrite performance for partition tables
Why is this PR needed?
With the increase in the number of overwrite loads for the partition table,
the time takes for each load keeps on increasing over time. This is because,

1. whenever a load overwrite for partition table is fired, it basically means
that we need to overwrite or drop the partitions if anything overlaps with
current partitions getting loaded. Since carbondata stores the partition
information in the segments file, to identify and drop partitions, it's
reading all the previous segment files to identify and drop the overwriting
partitions, which leads to a decrease in performance.

2. After partition load is completed, a cleanSegments method is called which
again reads segment file and table status file to identify MArked for delete
segments to clean. But Since the force clean is false and timeout also is
more than a day by default, it's not necessary to call this method.
Clean files should handle this part.

What changes were proposed in this PR?
1. we already have the information about current partitions, so with that first
identify if there are any partitions to overwrite, if yes then only we read segment
files to call dropParitition, else we don't read the segment files unnecessarily.
It also contains other refactoring to avoid reading table status file also.
2. no need to call clean segments after every load. Clean files will take care
to delete the expired ones.

This closes #4242