[CARBONDATA-4319] Fixed clean files not deleteting stale delete delta files after horizontal compaction
Why is this PR needed?
After horizontal compaction was performed on partition and non partition tables, the clean files
operation was not deleting the stale delete delta files. the code was removed as the part of clean
files refactoring done previously.

What changes were proposed in this PR?
Clean files with force option now handles removal of these stale delta files as well as the stale
tableupdatestatus file for both partition and non partition table.

This closes #4245