[CARBONDATA-4326] MV not hitting with multiple sessions issue fix
Why is this PR needed?
MV created in beeline not hitting in sql/shell and vice versa if both
beeline and sql/shell are running in parallel. Currently, If the view
catalog for a particular session is already initialized then the schemas
are not reloaded each time. So when mv is created in another session
and queried from the currently open session, mv is not hit.

What changes were proposed in this PR?
1.Reload mv catalog every time to getSchemas from the path. Register the
schema if not present in the catalog and deregister the schema if it's dropped.
2. When create SI is triggered, no need to try rewriting the plan and
check for mv schemas. So, returning plan if DeserializeToObject is present.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No, tested in cluster

This closes #4251