[CARBONDATA-4328] Load parquet table with options error message fix
Why is this PR needed?
If parquet table is created and load statement with options is
triggerred, then its failing with NoSuchTableException:
Table ${tableIdentifier.table} does not exist.

What changes were proposed in this PR?
As parquet table load is not handled, added a check to filter out
non-carbon tables in the parser. So that, the spark parser can handle the statement.

This closes #4253