[CARBONDATA-4339]Fix NullPointerException in load overwrite on partition table
Why is this PR needed?
After delete segment and clean files with force option true, the load overwrite
operation throws nullpointer exception. This is because when clean files with
force is done, except the 0th segment and last segment remaining marked for
delete will be moved to tablestatus.history file irrespective of the status of
the 0th and last segment. During overwrite load, the overwritten partition
will be dropped. Since all the segments are physically deleted with clean
files, and load model's load metadata details list contains 0th segment
which is marked for delete also leading to failure.

What changes were proposed in this PR?
When the valid segments are collected, filter using the segment's status to avoid the failure.

This closes #4280