[CARBONDATA-4347] Improve performance when delete empty partition directory
Why is this PR needed?
When delete multiple segments, carbondata tries to judge if partition directory is empty and delete it on each segments, but partiton path may be the same between two segments. So there is repetitive action.

What changes were proposed in this PR?
Collect a set of partition directories that are from segments to be deleted, judge empty and delete after all segments deleted.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No